feat: move More Actions from drawer to menu
feels like this makes the options easier to discover because they
have clear labels. matches Miro's more-menu style.

includes:
- removed nested menu item's overridden styles that were for matching
mui menu size - seems like they weren't necessary anymore (maybe
because of using `MenuItem` instead of `IconMenuItem`?)
- using `display: contents` so that we can have a parent div to
control show/hiding divider + more button together, while allowing
them to fit into the parent container's `flex` flow
- extracting `ScreenshotResolutionDialog` because it has a decent
amount of logic. later maybe we could consider making a
`MoreActionsMenu` directory where we could put each submenu as
its own file
- overriding some MUI sizing defaults for menu item / switch / radio
because the default for menu items feels too big. see comment in
theme.ts.
